Vorbereitung auf DataConsumer für ElasticSearch

+++ kopeme-core/src/test/java/de/dagere/kopeme/exampletests/pure/ExamplePurePerformanceTests.java
@@ -135,10 +135,10 @@ public class ExamplePurePerformanceTests {
 			@Override
 			public void checkValues(final TestResult tr) {
 				String CPUUSage = CPUUsageCollector.class.getName();
-				MatcherAssert.assertThat(tr.getValue(CPUUSage), Matchers.greaterThan(10L));
-				MatcherAssert.assertThat(tr.getValue(CPUUSage), Matchers.greaterThan((long) (tr.getHistoricalResults().getLastRunsAverage(CPUUSage, 5) * 0.80)));
+				MatcherAssert.assertThat(tr.getValue(CPUUSage), Matchers.greaterThan(10d));
+				MatcherAssert.assertThat(tr.getValue(CPUUSage), Matchers.greaterThan( (tr.getHistoricalResults().getLastRunsAverage(CPUUSage, 5) * 0.80)));
 				MatcherAssert.assertThat(tr.getValue(TimeDataCollector.class.getName()),
-						Matchers.lessThan((long) (tr.getHistoricalResults().getLastRunsAverage(TimeDataCollector.class.getName(), 5) * 1.30)));
+						Matchers.lessThan( (tr.getHistoricalResults().getLastRunsAverage(TimeDataCollector.class.getName(), 5) * 1.30)));
 			}
 		});
 	}
